refactor: move some code from ops to js.rs
